+++ Cmdr/BuiltInTypes/Type.lua
@@ -1,15 +1,15 @@
 local Util = require(script.Parent.Parent.Shared.Util)
 
 return function (cmdr)
-	local commandType = {
+	local typeType = {
 		Transform = function (text)
-			local findCommand = Util.MakeFuzzyFinder(cmdr:GetCommandNames())
+			local findCommand = Util.MakeFuzzyFinder(cmdr:GetTypeNames())
 
 			return findCommand(text)
 		end;
 
 		Validate = function (commands)
-			return #commands > 0, "No command with that name could be found."
+			return #commands > 0, "No type with that name could be found."
 		end;
 
 		Autocomplete = function (commands)
@@ -21,6 +21,6 @@ return function (cmdr)
 		end;
 	}
 
-	cmdr:RegisterType("command", commandType)
-	cmdr:RegisterType("commands", Util.MakeListableType(commandType))
+	cmdr:RegisterType("type", typeType)
+	cmdr:RegisterType("types", Util.MakeListableType(typeType))
 end