+++ Cmdr/Shared/Util.lua
@@ -468,4 +468,26 @@ function Util.EmulateTabstops(text, tabWidth)
 	return result
 end
 
+function Util.Mutex()
+	local queue = {}
+	local locked = false
+
+	return function ()
+		if locked then
+			table.insert(queue, coroutine.running())
+			coroutine.yield()
+		else
+			locked = true
+		end
+
+		return function()
+			if #queue > 0 then
+				coroutine.resume(table.remove(queue, 1))
+			else
+				locked = false
+			end
+		end
+	end
+end
+
 return Util