+++ Cmdr/BuiltInCommands/Utility/replace.lua
@@ -1,8 +1,12 @@
 return {
 	Name = "replace";
-	Aliases = {};
+	Aliases = {"gsub", "//"};
 	Description = "Replaces text A with text B";
 	Group = "DefaultUtil";
+	AutoExec = {
+		"alias \"map|Maps a CSV into another CSV\" replace $1{string|CSV} ([^,]+) \"$2{string|mapped value|Use %1 to insert the element}\"",
+		"alias join replace $1{string|CSV} , $2{string|Delimiter}"
+	},
 	Args = {
 		{
 			Type = "string";
@@ -17,7 +21,8 @@ return {
 		{
 			Type = "string";
 			Name = "Replacement";
-			Description = "The string to replace matches (%1 to insert matches)."
+			Description = "The string to replace matches (%1 to insert matches).";
+			Default = "";
 		},
 	};